scripts/requirements.txt
# Article Ingestion Pipeline Dependencies

# Claude API
anthropic>=0.39.0

# HTTP requests
httpx>=0.27.0

# HTML parsing
beautifulsoup4>=4.12.0

# Vector database
faiss-cpu>=1.8.0

# Embeddings
sentence-transformers>=2.2.0

# Utilities
numpy>=1.24.0
